Crate stdlib

Constants

COMPILED_EXTENSION
The extension for compiled files

COMPILED_OUTPUT_PATH
The output path under which compiled files will be put

COMPILED_STDLIB_NAME
The file name for the compiled stdlib

COMPILED_TRANSACTION_SCRIPTS_ABI_DIR
The output path for transaction script ABIs.

COMPILED_TRANSACTION_SCRIPTS_DIR
The output path under which compiled script files can be found

DEBUG_MODULE_FILE_NAME
The file name of the debug module

MOVE_EXTENSION

STD_LIB_DIR

STD_LIB_DOC_DIR
The output path for stdlib documentation.

TRANSACTION_BUILDERS_GENERATED_SOURCE_PATH
Where to write generated transaction builders.

TRANSACTION_SCRIPTS

TRANSACTION_SCRIPTS_DOC_DIR
The output path for transaction script documentation.

Functions

build_stdlib

build_stdlib_doc

build_transaction_script_abi

build_transaction_script_doc

compile_script

filter_move_files

generate_rust_transaction_builders

save_binary

stdlib_files