Function stdlib::stdlib_modules

pub fn stdlib_modules(option: StdLibOptions) -> &'static [VerifiedModule]

Returns a reference to the standard library. Depending upon the option flag passed in either a staged version of the standard library will be returned or a new freshly built stdlib will be used.