Module stdlib::transaction_scripts

Structs

BLOCK_PROLOGUE_TXN_BODY

CREATE_ACCOUNT_TXN_BODY

MINT_TXN_BODY

PEER_TO_PEER_TRANSFER_TXN_BODY

ROTATE_AUTHENTICATION_KEY_TXN_BODY

ROTATE_CONSENSUS_PUBKEY_TXN_BODY

Functions

block_prologue
Returns the source code for the block prologue script

create_account
Returns the source code for create-account transaction script.

mint
Returns the source code for the mint transaction script.

peer_to_peer
Returns the source code for peer-to-peer transaction script.

rotate_consensus_pubkey
Returns the source code for the rotate-consensus-pubkey script.

rotate_key
Returns the source code for the rotate-key transaction script.