Enum stdlib::transaction_scripts::StdlibScript

pub enum StdlibScript {
    AddValidator,
    AddCurrency,
    AddCurrencyToAccount,
    AllowChildAccounts,
    ApplyForAssociationAddress,
    ApplyForAssociationPrivilege,
    ApplyForChildVaspCredential,
    ApplyForParentCapability,
    ApplyForRootVasp,
    ApprovedPayment,
    Burn,
    CancelBurn,
    CreateAccount,
    CreateEmptyAccount,
    EmptyScript,
    GrantAssociationAddress,
    GrantAssociationPrivilege,
    GrantChildAccount,
    GrantParentAccount,
    GrantVaspAccount,
    Mint,
    MintLbr,
    ModifyPublishingOption,
    PeerToPeer,
    PeerToPeerWithMetadata,
    Preburn,
    PublishSharedEd2551PublicKey,
    RecertifyChildAccount,
    RegisterApprovedPayment,
    RegisterPreburner,
    RegisterValidator,
    RemoveAssociationAddress,
    RemoveAssociationPrivilege,
    RemoveChildAccount,
    RemoveParentAccount,
    RemoveValidator,
    RotateAuthenticationKey,
    RotateCompliancePublicKey,
    RotateConsensusPubkey,
    RotateSharedEd2551PublicKey,
    UnmintLbr,
    UpdateLibraVersion,
    UpdateExchangeRate,
    UpdateMintingAbility,
}

All of the Move transaction scripts that can be executed on the Libra blockchain

Variants

AddValidator

AddCurrency

AddCurrencyToAccount

AllowChildAccounts

ApplyForAssociationAddress

ApplyForAssociationPrivilege

ApplyForChildVaspCredential

ApplyForParentCapability

ApplyForRootVasp

ApprovedPayment

Burn

CancelBurn

CreateAccount

CreateEmptyAccount

EmptyScript

GrantAssociationAddress

GrantAssociationPrivilege

GrantChildAccount

GrantParentAccount

GrantVaspAccount

Mint

MintLbr

ModifyPublishingOption

PeerToPeer

PeerToPeerWithMetadata

Preburn

PublishSharedEd2551PublicKey

RecertifyChildAccount

RegisterApprovedPayment

RegisterPreburner

RegisterValidator

RemoveAssociationAddress

RemoveAssociationPrivilege

RemoveChildAccount

RemoveParentAccount

RemoveValidator

RotateAuthenticationKey

RotateCompliancePublicKey

RotateConsensusPubkey

RotateSharedEd2551PublicKey

UnmintLbr

UpdateLibraVersion

UpdateExchangeRate

UpdateMintingAbility

Methods

impl StdlibScript

pub fn all() -> Vec<Self>

Return a vector containing all of the standard library scripts (i.e., all inhabitants of the StdlibScript enum)

pub fn whitelist() -> Vec<[u8; 32]>

Construct the whitelist of script hashes used to determine whether a transaction script can be executed on the Libra blockchain

pub fn name(self) -> String

Return a lowercase-underscore style name for this script

pub fn is(code_bytes: &[u8]) -> bool

Return true if code_bytes is the bytecode of one of the standard library scripts

pub fn compiled_bytes(self) -> CompiledBytes

Return the Move bytecode produced by compiling this script. This will almost always read from disk rather invoking the compiler; genesis is the only exception.

pub fn hash(self) -> HashValue

Return the sha3-256 hash of the compiled script bytes