Crate stdlib

Modules

utils

Constants

COMPILED_EXTENSION
The extension for compiled files

COMPILED_OUTPUT_PATH
The output path under which compiled files will be put

COMPILED_STDLIB_DIR
The file name for the compiled stdlib

COMPILED_TRANSACTION_SCRIPTS_ABI_DIR
The output path for transaction script ABIs.

COMPILED_TRANSACTION_SCRIPTS_DIR
The output path under which compiled script files can be found

DEBUG_MODULE_FILE_NAME
The file name of the debug module

ERROR_DESC_DIR

ERROR_DESC_EXTENSION

ERROR_DESC_FILENAME

MOVE_EXTENSION

PACKED_TYPES_DIR

PACKED_TYPES_EXTENSION

PACKED_TYPES_FILENAME

REFERENCES_DOC_TEMPLATE
Path to the references template.

SPEC_DOC_TEMPLATE
The specification root template for scripts and stdlib.

STD_LIB_DIR

STD_LIB_DOC_DIR
The output path for stdlib documentation.

STD_LIB_DOC_TEMPLATE
The documentation root template for stdlib.

TRANSACTION_BUILDERS_GENERATED_SOURCE_PATH
Where to write generated transaction builders.

TRANSACTION_SCRIPTS

TRANSACTION_SCRIPTS_DOC_DIR
The output path for transaction script documentation.

TRANSACTION_SCRIPT_DOC_TEMPLATE
The documentation root template for scripts.

Functions

build_stdlib

build_stdlib_doc

build_stdlib_error_code_map

build_transaction_script_abi

build_transaction_script_doc

compile_script

filter_move_bytecode_files

filter_move_files

generate_rust_transaction_builders

get_packed_types_path

path_in_crate

save_binary

script_files

stdlib_bytecode_files

stdlib_files